added data into entity file

+++ exam-api/src/questions/questions.service.ts
@@ -11,6 +11,10 @@ export class QuestionsService {
         option: createQuestionDto.options,
         question_time: createQuestionDto?.question_time,
         status: createQuestionDto?.status,
+        level_id: createQuestionDto?.level_id,
+        module_id: createQuestionDto?.module_id,
+        marks: createQuestionDto?.marks,
+        option_type: createQuestionDto?.option_type,
       },
     });
 
@@ -41,6 +45,11 @@ export class QuestionsService {
   }
 
   async update(id: string, updateRestApiDto: any) {
+    const find = await this.prisma.questions.findUnique({ where: { id: id } });
+    if (!find) {
+      return 'data is not exist!';
+    }
+
     const updatedOptions = await this.prisma.questions.update({
       where: {
         id: id,
@@ -54,6 +63,12 @@ export class QuestionsService {
   }
 
   async remove(idd: string) {
+    const find_del = await this.prisma.questions.findUnique({
+      where: { id: idd },
+    });
+    if (!find_del) {
+      return 'data is not exist!';
+    }
     const delete_question = await this.prisma.questions.delete({
       where: {
         id: idd,
